% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qfasa.R
\name{AIT.dist}
\alias{AIT.dist}
\title{Computes the difference between to vectors of compositional data
as described in aitchison (1992) "measures of compositional difference".
Note that this function is different to \code{compdiff()} in that it takes square root.
\code{compdiff()} is used in calculating diet estimates.}
\usage{
AIT.dist(x, bigX)
}
\arguments{
\item{x}{TODO}

\item{bigX}{TODO}
}
\description{
Computes the difference between to vectors of compositional data
as described in aitchison (1992) "measures of compositional difference".
Note that this function is different to \code{compdiff()} in that it takes square root.
\code{compdiff()} is used in calculating diet estimates.
}

