% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qfasa.R
\name{AIT.more}
\alias{AIT.more}
\title{Used to provide additional information on model components when
alpha corresponds to the QFASA diet estimates (i.e. estimates
that minimized the ait distance). Used in \code{solnp()} as the
objective function to be minimized.}
\usage{
AIT.more(alpha, seal, prey.quantiles)
}
\arguments{
\item{alpha}{vector over which minimization takes place}

\item{seal}{vector of  fatty acid compositions of seal}

\item{prey.quantiles}{matrix of fatty acid composition of
prey. Each row contains an individual prey from a different
species.}
}
\description{
Used to provide additional information on model components when
alpha corresponds to the QFASA diet estimates (i.e. estimates
that minimized the ait distance). Used in \code{solnp()} as the
objective function to be minimized.
}

