% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qfasa.R
\name{AIT.obj}
\alias{AIT.obj}
\title{Similar to \code{optcompdiff.obj()} but does not normalize alpha.
Used in \code{solnp()} as the objective function to be minimized.}
\usage{
AIT.obj(alpha, seal, prey.quantiles)
}
\arguments{
\item{alpha}{vector over which minimization takes place.}

\item{seal}{vector of  fatty acid compositions of seal.}

\item{prey.quantiles}{matrix of fatty acid composition of
prey. Each row contains an individual prey from a different
species.}
}
\description{
Similar to \code{optcompdiff.obj()} but does not normalize alpha.
Used in \code{solnp()} as the objective function to be minimized.
}

