% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/All_Selection_Algorithm_Code.R
\name{bestBackwardModel}
\alias{bestBackwardModel}
\title{Finds the species combination with the best (lowest) IC value achievable by
dropping a single species from the prey.mat. Used for each iteration of
backward elimination.}
\usage{
bestBackwardModel(
  pred.mat,
  prey.mat,
  cal.vec,
  fat.vec,
  ext.fa,
  starting.estimates,
  k = 2,
  cutoff = 0.01,
  silence = FALSE
)
}
\description{
Finds the species combination with the best (lowest) IC value achievable by
dropping a single species from the prey.mat. Used for each iteration of
backward elimination.
}
\keyword{internal}
