% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/All_Selection_Algorithm_Code.R
\name{bestForwardModel}
\alias{bestForwardModel}
\title{Finds the species combination with the best# (lowest) IC value achievable by
adding a single species from full.prey.mat. Used for each iteration of
forward selection.}
\usage{
bestForwardModel(
  pred.mat,
  full.prey.mat,
  prey.mat,
  cal.vec,
  fat.vec,
  ext.fa,
  k = 2,
  silence = FALSE
)
}
\description{
Finds the species combination with the best# (lowest) IC value achievable by
adding a single species from full.prey.mat. Used for each iteration of
forward selection.
}
\keyword{internal}
