% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/All_Selection_Algorithm_Code.R
\name{bestSubset}
\alias{bestSubset}
\title{Find the pair of starting species with the best (lowest) IC value among all
possible pairs of species in prey.mat Used in forward selection when no
starting species are specified}
\usage{
bestSubset(
  pred.mat,
  prey.mat,
  cal.vec,
  fat.vec,
  ext.fa,
  k = 2,
  silence = FALSE
)
}
\description{
Find the pair of starting species with the best (lowest) IC value among all
possible pairs of species in prey.mat Used in forward selection when no
starting species are specified
}
\keyword{internal}
