% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Confidence_Interval_Beta_Distribution.R
\name{beta.pval.k}
\alias{beta.pval.k}
\title{Calculate p-value of a given prey type diet proportion under the
predator diets and estimated diet distribution provided.}
\usage{
beta.pval.k(par.list, R, diet.test.k, p.mat, k)
}
\arguments{
\item{par.list}{a list of R.p lists of I beta distribution parameters phi
and theta that define diet proportion estimates for each of the
prey species. Effectively R.p beta distibutions for each of the
I prey species which we bootstrap to calculate p-values.}

\item{R}{number of bootstrap replicates to use in p-value estimation.}

\item{diet.test.k}{diet proportion for which we want the p-value}

\item{p.mat}{predator diet estimates.}

\item{k}{prey species index 1..I}
}
\value{
p-value p-value
}
\description{
Calculate p-value of a given prey type diet proportion under the
predator diets and estimated diet distribution provided.
}
\keyword{internal}
