% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/All_CI_Code.R
\name{bias.comp}
\alias{bias.comp}
\title{Calculate bias correction for confidence intervals.}
\usage{
bias.comp(diet.est, prey.mat, R.bias, dist.meas, ext.fa, gamma = 1)
}
\arguments{
\item{diet.est}{predator diet estimates}

\item{prey.mat}{matrix of prey FA signatures}

\item{R.bias}{number of replicates}

\item{dist.meas}{distance measure}

\item{ext.fa}{subset of FA's to use.}
}
\value{
estimate of bias for each prey type
}
\description{
Calculate bias correction for confidence intervals.
}
\keyword{internal}
