% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CC.R
\docType{data}
\name{CC}
\alias{CC}
\title{Fatty acid calibration coefficients.}
\format{
A data frame with 66 observations and 2 variables:
\describe{
\item{FA}{fatty acid names}
\item{CC}{calibration coefficient for corresponding fatty acid}
}
}
\usage{
CC
}
\description{
Fatty acid calibration coefficients.
}
\keyword{datasets}
