% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qfasa.R
\name{chisq.CA}
\alias{chisq.CA}
\title{Called by create.d.mat to compute the chisquare distance}
\usage{
chisq.CA(x1, x2)
}
\arguments{
\item{x1}{vector}

\item{x2}{vector}
}
\description{
Called by create.d.mat to compute the chisquare distance
}
