% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qfasa.R
\name{chisq.dist}
\alias{chisq.dist}
\title{Calculates CS distance as described in testing for diff in diet paper.
This is different to chisq.ca in directory run.oct11.r.1 because
x1 and x2 are not assumed to be transformed.}
\usage{
chisq.dist(x.1, x.2, alpha)
}
\arguments{
\item{x.1}{vector}

\item{x.2}{vector}

\item{alpha}{TODO}
}
\description{
Calculates CS distance as described in testing for diff in diet paper.
This is different to chisq.ca in directory run.oct11.r.1 because
x1 and x2 are not assumed to be transformed.
}

