% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/All_CI_Code.R
\name{comp.beta.pval.k}
\alias{comp.beta.pval.k}
\title{Calculate p-value corresponding to a specified null value using bootstrapping}
\usage{
comp.beta.pval.k(par.list, R, diet.test.k, p.mat, k)
}
\arguments{
\item{par.list}{a list of R.p lists of I beta distribution parameters phi
and theta that define diet proportion estimates for each of the
prey species. Effectively R.p beta distibutions for each of the
I prey species which we bootstrap to calculate p-values.}

\item{R}{number of bootstrap replicates to use in p-value estimation.}

\item{diet.test.k}{null value}

\item{p.mat}{predator diet estimates.}

\item{k}{prey species index 1..I}
}
\value{
p-value p-value
}
\description{
Calculate p-value corresponding to a specified null value using bootstrapping
}
\keyword{internal}
