% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/All_CI_Code.R
\name{comp.Tstar.beta}
\alias{comp.Tstar.beta}
\title{Generate bootstrap replicates of diet proportion estimates for
given prey species}
\usage{
comp.Tstar.beta(p.k, theta.boot, mu.null, phi.boot, R.boot)
}
\arguments{
\item{p.k}{list of predator diet proportions for prey species k}

\item{theta.boot}{distribution parameter}

\item{mu.null}{distribution parameter}

\item{phi.boot}{distribution parameter}

\item{R.boot}{number of bootstrap replicates to create}
}
\value{
list of diet proportion replicates
}
\description{
Generate bootstrap replicates of diet proportion estimates for
given prey species
}
\keyword{internal}
