% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qfasa.R
\name{CS.obj}
\alias{CS.obj}
\title{Similar to AIT.obj and KL.obj but does not require changing zeros.
Used in \code{solnp()} as the objective function to be minimized.}
\usage{
CS.obj(alpha, seal, prey.quantiles, gamma)
}
\arguments{
\item{alpha}{vector over which minimization takes place.}

\item{seal}{vector of  fatty acid compositions of seal.}

\item{prey.quantiles}{matrix of fatty acid composition of
prey. Each row contains an individual prey from a different
species.}

\item{gamma}{parameter passed to chisq.dist}
}
\description{
Similar to AIT.obj and KL.obj but does not require changing zeros.
Used in \code{solnp()} as the objective function to be minimized.
}

