% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/All_Rep_Code.R
\name{cs_distance.mat}
\alias{cs_distance.mat}
\title{CALCULATE CS DISTANCE BETWEEN EACH TWO PREDATORS IN dataset}
\usage{
cs_distance.mat(dataset, alpha = 1)
}
\description{
CALCULATE CS DISTANCE BETWEEN EACH TWO PREDATORS IN dataset
}
\keyword{internal}
