% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/All_CI_Code.R,
%   R/Confidence_Interval_Beta_Distribution.R
\name{data.sim.beta}
\alias{data.sim.beta}
\title{Bootstrap ran.gen function:}
\usage{
data.sim.beta(data, mle)

data.sim.beta(data, mle)
}
\description{
Bootstrap ran.gen function:

Bootstrap ran.gen function:
}
\keyword{internal}
