% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/All_Selection_Algorithm_Code.R
\name{evaluateModel}
\alias{evaluateModel}
\title{Get nll and IC values for a given model (note that the model is determined by
the species included in prey.mat - prey.mat is adjusted to add or remove
species before entering this function}
\usage{
evaluateModel(pred.mat, prey.mat, cal.vec, fat.vec, ext.fa, k = 2)
}
\description{
Get nll and IC values for a given model (note that the model is determined by
the species included in prey.mat - prey.mat is adjusted to add or remove
species before entering this function
}
\keyword{internal}
