% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FAset.R
\docType{data}
\name{FAset}
\alias{FAset}
\title{List of fatty acids used in sample prey and predator data sets,
\code{preyFAs} and \code{predatorFAs} respectively.}
\format{
A data frame with 39 observations and 1 variable:
\describe{
\item{FA}{Fatty acid name}
}
}
\usage{
FAset
}
\description{
List of fatty acids used in sample prey and predator data sets,
\code{preyFAs} and \code{predatorFAs} respectively.
}
\keyword{datasets}
