% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Confidence_Interval_Beta_Distribution.R
\name{gen.pseudo.seals}
\alias{gen.pseudo.seals}
\title{Generate pseudo predators with ith predator having true diet
given by ith row of diet.null.mat.}
\usage{
gen.pseudo.seals(prey.mat, diet.null.mat, cal.mat, fat.cont, noise, nprey)
}
\arguments{
\item{prey.mat}{matrix containing a representative FA signature
from each prey group (usually the mean). The first column must
index the prey group.}

\item{diet.null.mat}{null diet}

\item{cal.mat}{matrix of calibration coefficients where the \emph{i} th
column is to be used with the \emph{i} th predator.}

\item{fat.cont}{vector of fat content}

\item{noise}{proportion of noise to add to the simulation.}

\item{nprey}{number of prey to sample.}
}
\description{
Generate pseudo predators with ith predator having true diet
given by ith row of diet.null.mat.
}
