% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qfasa.R
\name{KL.dist}
\alias{KL.dist}
\title{Similar to KL but requires two vectors as input.}
\usage{
KL.dist(x, y)
}
\arguments{
\item{x}{vector}

\item{y}{vector}
}
\value{
Kulback-Liebler distance between x and y.
}
\description{
Similar to KL but requires two vectors as input.
}

