% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qfasa.R
\name{KL.more}
\alias{KL.more}
\title{Used to provide additional information on various model components
evaluated at the optimal solution i.e. using the QFASA diet estimates and
Kullback-Leibler distance measure.}
\usage{
KL.more(alpha, seal, prey.quantiles)
}
\arguments{
\item{alpha}{compositional QFASA diet estimate.}

\item{seal}{fatty acid signature of predator.}

\item{prey.quantiles}{matrix of fatty acid signatures of
prey. Each row contains an individual prey signature from a different
species.}
}
\description{
Used to provide additional information on various model components
evaluated at the optimal solution i.e. using the QFASA diet estimates and
Kullback-Leibler distance measure.
}

