% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qfasa.R
\name{KL.obj}
\alias{KL.obj}
\title{The objective function to be minimized by code{solnp()}. Similar to
optquantile.obj2 but does not normalize alpha.}
\usage{
KL.obj(alpha, seal, prey.quantiles)
}
\arguments{
\item{alpha}{vector over which minimization takes place.}

\item{seal}{vector of fatty acid compositions of seal.}

\item{prey.quantiles}{matrix of fatty acid composition of
prey. Each row contains an individual prey from a different species.}
}
\description{
The objective function to be minimized by code{solnp()}. Similar to
optquantile.obj2 but does not normalize alpha.
}

