% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/All_Selection_Algorithm_Code.R
\name{likelihoodEstimates}
\alias{likelihoodEstimates}
\title{Get simplified MLE estimates}
\usage{
likelihoodEstimates(pred.mat, prey.mat, cal.mat, fat.vec, ext.fa)
}
\description{
Get simplified MLE estimates
}
\keyword{internal}
