% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qfasa.R
\name{mean.geometric}
\alias{mean.geometric}
\title{Returns the geometric mean of a compositional vector}
\usage{
\method{mean}{geometric}(x)
}
\arguments{
\item{x}{compositional vector}
}
\description{
Returns the geometric mean of a compositional vector
}
