% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qfasa.R
\name{mean_geometric}
\alias{mean_geometric}
\title{Returns the geometric mean of a compositional vector}
\usage{
mean_geometric(x)
}
\arguments{
\item{x}{compositional vector}
}
\description{
Returns the geometric mean of a compositional vector
}
