% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qfasa.R
\name{MEANmeth}
\alias{MEANmeth}
\title{Returns the multivariate mean FA signature of each prey group
entered into the QFASA model. Result can be passed to prey.mat in
\code{p.QFASA()}.}
\usage{
MEANmeth(prey.mat)
}
\arguments{
\item{prey.mat}{matrix containing the FA signatures of the
prey. The first column indexes the prey group.}
}
\description{
Returns the multivariate mean FA signature of each prey group
entered into the QFASA model. Result can be passed to prey.mat in
\code{p.QFASA()}.
}

