% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/All_MUFASA_Code.R
\name{mod.zeros.FA.sig.mat}
\alias{mod.zeros.FA.sig.mat}
\title{Modifies the zeros for a sample of FA signatures using the multiplicative
replacement method (and same delta for every zero).}
\usage{
mod.zeros.FA.sig.mat(y.mat, delta)
}
\description{
Modifies the zeros for a sample of FA signatures using the multiplicative
replacement method (and same delta for every zero).
}
\keyword{internal}
