% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/All_MUFASA_Code.R
\name{mod.zeros.FA.sig}
\alias{mod.zeros.FA.sig}
\title{Modifies the zeros for a single FA signature using the multiplicative
replacement method (and same delta for every zero).}
\usage{
mod.zeros.FA.sig(y, delta)
}
\description{
Modifies the zeros for a single FA signature using the multiplicative
replacement method (and same delta for every zero).
}
\keyword{internal}
