% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/All_Simplified_MUFASA_Code.R
\name{multiplicativeReplacement}
\alias{multiplicativeReplacement}
\title{Multiplicative replacement of zeroes}
\usage{
multiplicativeReplacement(compositional.mat, delta = 1e-05)
}
\arguments{
\item{compositional.mat}{matrix containing compositional data (for example, FA signatures)
that may contain zeros.}

\item{delta}{imputed value}
}
\value{
The compositional matrix with zeros modified.
}
\description{
Multiplicative replacement of zeroes
}
