% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qfasa.R
\name{multiplot}
\alias{multiplot}
\title{Multiplot}
\usage{
multiplot(..., plotlist = NULL, file, cols = 1, layout = NULL)
}
\arguments{
\item{...}{TODO}

\item{plotlist}{TODO}

\item{file}{TODO}

\item{cols}{TODO}

\item{layout}{TODO}
}
\description{
Multiplot
}
