% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Confidence_Interval_Beta_Distribution.R
\name{p.beta}
\alias{p.beta}
\title{Bootstrap statistic function: in this case it is the mean
(meanBEZI) of the fitted (gamlss) ZIB distribution.}
\usage{
p.beta(data)
}
\arguments{
\item{data}{data}
}
\value{
the expected value of the response for a fitted model
}
\description{
Bootstrap statistic function: in this case it is the mean
(meanBEZI) of the fitted (gamlss) ZIB distribution.
}
\keyword{internal}
