% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/All_MUFASA_Code.R
\name{POOLVARmeth}
\alias{POOLVARmeth}
\title{Computes within species variance-covariance matrices on transformed scaled,
along wtih a pooled estimate.}
\usage{
POOLVARmeth(prey.mat)
}
\arguments{
\item{prey.mat}{matrix containing transformed FA signatures of the prey.
Note that the first column indexes prey type.}
}
\value{
Returns the variance-covariance matrix of each prey type as well as
        a pooled estimate of the variance-covariance matrix
}
\description{
Computes within species variance-covariance matrices on transformed scaled,
along wtih a pooled estimate.
}
