% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prey.on.prey.R
\name{prey.on.prey}
\alias{prey.on.prey}
\title{Each prey fatty acid signature is systematically removed from the
supplied prey database and its QFASA diet estimate is obtained by
treating the individual as a predator.}
\usage{
prey.on.prey(preybase, dist.meas, gamma = 1)
}
\arguments{
\item{preybase}{first column is name of species and remaining columns are fatty acids.}

\item{dist.meas}{see help file for \code{\link{p.QFASA}}.}

\item{gamma}{see help file for \code{\link{p.QFASA}}.}
}
\value{
diet estimate
}
\description{
Each prey fatty acid signature is systematically removed from the
supplied prey database and its QFASA diet estimate is obtained by
treating the individual as a predator.
}
\examples{
data(preyFAs)
my.preybase <- preyFAs[, -c(1,3)]

# Note: uncomment examples to run. CRAN tests fail because execution time > 5 seconds
# diets.out <- prey.on.prey(my.preybase, 2)
# round(MEANmeth(diets.out), 3)
     
}
