% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preyFAs.R
\docType{data}
\name{preyFAs}
\alias{preyFAs}
\title{Prey fatty acid signatures. Each prey signature is a row with fatty
acid proportions in columns.}
\format{
A data frame with 302 observations and 70 variables:
\describe{
\item{Lab.Code}{TODO}
\item{Species}{TODO}
\item{lipid}{TODO}
\item{c12.0}{}
\item{c13.0}{}
\item{Iso14}{}
\item{c14.0}{}
\item{c14.1w9}{}
\item{c14.1w7}{}
\item{c14.1w5}{}
\item{Iso15}{}
\item{Anti15}{}
\item{c15.0}{}
\item{c15.1w8}{}
\item{c15.1w6}{}
\item{Iso16}{}
\item{c16.0}{}
\item{c16.1w11}{}
\item{c16.1w9}{}
\item{c16.1w7}{}
\item{c7Me16.0}{}
\item{c16.1w5}{}
\item{c16.2w6}{}
\item{Iso17}{}
\item{c16.2w4}{}
\item{c16.3w6}{}
\item{c17.0}{}
\item{c16.3w4}{}
\item{c17.1}{}
\item{c16.3w1}{}
\item{c16.4w3}{}
\item{c16.4w1}{}
\item{c18.0}{}
\item{c18.1w13}{}
\item{c18.1w11}{}
\item{c18.1w9}{}
\item{c18.1w7}{}
\item{c18.1w5}{}
\item{c18.2d5.11}{}
\item{c18.2w7}{}
\item{c18.2w6}{}
\item{c18.2w4}{}
\item{c18.3w6}{}
\item{c18.3w4}{}
\item{c18.3w3}{}
\item{c18.3w1}{}
\item{c18.4w3}{}
\item{c18.4w1}{}
\item{c20.0}{}
\item{c20.1w11}{}
\item{c20.1w9}{}
\item{c20.1w7}{}
\item{c20.2w9}{}
\item{c20.2w6}{}
\item{c20.3w6}{}
\item{c20.4w6}{}
\item{c20.3w3}{}
\item{c20.4w3}{}
\item{c20.5w3}{}
\item{c22.1w11}{}
\item{c22.1w9}{}
\item{c22.1w7}{}
\item{c22.2w6}{}
\item{c21.5w3}{}
\item{c22.4w6}{}
\item{c22.5w6}{}
\item{c22.4w3}{}
\item{c22.5w3}{}
\item{c22.6w3}{}
\item{c24.1w9}{}
}
}
\usage{
preyFAs
}
\description{
The prey file should contain all of the individual fatty acid signatures of
the prey and their lipid contents (where appropriate) - a matrix of
the mean values for the FAs (prey.matrix) by the designated prey
modelling group is then calculated using the MEANmeth function.
}
\details{
Like the predator .csv file you can have as many tombstone data
columns as required but there must be at least one column that
identifies the modelling group, in this case, Species.

Unlike the predator data, the prey data is not subsetted and
renomalized during the modelling so the prey file needs to be
subsetted for the desired fatty acid set and renormalized to
sum to 1 prior to calculating the mean values.

The full FA set is extracted from the data frame
(columns 4 onward), subsetted for the FA set in use and then
renormalized over 1. The modelling group names (the "Species" column
in this case) is then added back to the subsetted and renormalized
data (as the first column) and the average values calculated using
the MEANmeth function. Note that for the MEANmeth function to work
the modelling group name must be in the first column.
}
\keyword{datasets}
