% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/All_Rep_Code.R
\name{pseudo.pred.rep}
\alias{pseudo.pred.rep}
\title{FUNCTION TO GENERATE ns PSEUDO SEALS WHERE ns IS THE NUMBER OF ROWS IN
\code{diet.mat.seal} AND RETURNS THE CORRESPONDING QFASA DIET ESTIMATES.
ASSUMES EACH ROW IS A DIET VECTOR AND EACH COLUMN CORRESPONDS TO 1 PREY
SPECIES IN THE DIET}
\usage{
pseudo.pred.rep(
  diet.mat.seal,
  prey.database,
  fatcont.mat,
  dist.meas,
  ext.fa,
  gamma = 1
)
}
\description{
FUNCTION TO GENERATE ns PSEUDO SEALS WHERE ns IS THE NUMBER OF ROWS IN
\code{diet.mat.seal} AND RETURNS THE CORRESPONDING QFASA DIET ESTIMATES.
ASSUMES EACH ROW IS A DIET VECTOR AND EACH COLUMN CORRESPONDS TO 1 PREY
SPECIES IN THE DIET
}
\keyword{internal}
