% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/All_Rep_Code.R
\name{pseudo.pred.table}
\alias{pseudo.pred.table}
\title{FUNCTION TO GENERATE ns SEALS FOR EACH OF ny YEARS}
\usage{
pseudo.pred.table(
  diet.rep.mat,
  prey.database,
  fatcont.mat,
  dist.meas,
  ext.fa,
  gamma = 1
)
}
\description{
FUNCTION TO GENERATE ns SEALS FOR EACH OF ny YEARS
}
\keyword{internal}
