% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/All_CI_Code.R,
%   R/Confidence_Interval_Beta_Distribution.R
\name{pseudo.seal}
\alias{pseudo.seal}
\title{Generate a single pseudo predator FA signature}
\usage{
pseudo.seal(prey.sim, diet, noise, nprey, cal, fat.cont, specify.noise)

pseudo.seal(prey.sim, diet, noise, nprey, cal, fat.cont, specify.noise)
}
\arguments{
\item{prey.sim}{OUTPUT OF split.prey}

\item{diet}{DIET COMPOSITION VECTOR  (NOTE: THIS VECTOR SHOULD SUM TO 1-NOISE.  THE NOISE WILL BE ADDED TO THE diet VECTOR.)}

\item{noise}{AMOUNT OF NOISE}

\item{nprey}{nprey TOTAL NUMBER OF PREY TO BE SAMPLED}

\item{cal}{CALIBRATION FACTORS}

\item{fat.cont}{VECTOR OF FAT CONTENT OF LENGTH=I (# OF SPECIES)}

\item{specify.noise}{A BOOLEAN VECTOR WITH TRUES DENOTING SPECIES TO USE IN NOISE.}
}
\value{
seal.star SIMULATED SEAL FA SIGNATURE.

seal.star SIMULATED SEAL FA SIGNATURE.
}
\description{
THIS IS THE NEW pseudo.seal FUNCTION THAT ALLOWS 1) FAT CONTENT
TO BE INCLUDED IN THE GENERATED SEALS AND 2) SOME SPECIES TO BE
TRULY ZERO (THAT IS,"ZERO SPECIES" DO NOT HAVE TO BE INCLUDED IN THE
"NOISE" )
NOTE:  IT IS ASSUMED THAT SUM(DIET) IS 1-NOISE

THIS IS THE NEW pseudo.seal FUNCTION THAT ALLOWS 1) FAT CONTENT
TO BE INCLUDED IN THE GENERATED SEALS AND 2) SOME SPECIES TO BE
TRULY ZERO (THAT IS,"ZERO SPECIES" DO NOT HAVE TO BE INCLUDED IN THE
"NOISE" )
NOTE:  IT IS ASSUMED THAT SUM(DIET) IS 1-NOISE
}
\keyword{internal}
