% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qfasa.R
\name{QFASA.const.eqn}
\alias{QFASA.const.eqn}
\title{QFASA.const.eqn}
\usage{
QFASA.const.eqn(alpha, seal, prey.quantiles, gamma)
}
\arguments{
\item{alpha}{TODO}

\item{seal}{TODO}

\item{prey.quantiles}{TODO}

\item{gamma}{TODO}
}
\description{
QFASA.const.eqn
}

