% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qfasa.R
\docType{package}
\name{QFASA}
\alias{QFASA}
\alias{QFASA-package}
\title{QFASA: A package for Quantitative Fatty Acid Signature Analysis}
\description{
Accurate estimates of the diets of predators are required
in many areas of ecology, but for many species current methods are
imprecise, limited to the last meal, and often biased. The diversity
of fatty acids and their patterns in organisms, coupled with the
narrow limitations on their biosynthesis, properties of digestion in
monogastric animals, and the prevalence of large storage reservoirs of
lipid in many predators, led us to propose the use of quantitative
fatty acid signature analysis (QFASA) to study predator diets.
}

