% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/All_CI_Code.R
\name{root.beta}
\alias{root.beta}
\title{Find root (i.e. solve f(pio) = PVAL(pio) = alpha1 using either uniroot or optimize.}
\usage{
root.beta(x1, x2, alpha, par.list, R, p.mat, k, low = TRUE)
}
\arguments{
\item{x1}{lower starting value}

\item{x2}{upper starting value}

\item{alpha}{simultaneous confidence level}

\item{par.list}{parameters needed to compute p-value by bootstrapping}

\item{R}{number of bootstraps}

\item{p.mat}{predator diet estimates}

\item{k}{prey species index 1...I}

\item{low}{TRUE if computing lower interval.}
}
\value{
root which will be lower or upper CI limit
}
\description{
Find root (i.e. solve f(pio) = PVAL(pio) = alpha1 using either uniroot or optimize.
}
\keyword{internal}
