% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qfasa.R
\name{testfordiff.ind.boot.fun}
\alias{testfordiff.ind.boot.fun}
\title{Called by \code{testfordiff.ind.boot()}.}
\usage{
testfordiff.ind.boot.fun(data, i, ns1, change.zero = 1e-05)
}
\arguments{
\item{data}{sample of compositional data}

\item{i}{}

\item{ns1}{sample size of compdata.1}

\item{change.zero}{}
}
\description{
Called by \code{testfordiff.ind.boot()}.
}
