% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Confidence_Interval_Beta_Distribution.R
\name{Tstar.beta}
\alias{Tstar.beta}
\title{Generate bootstrap replicates of diet proportion estimates for
given prey species}
\usage{
Tstar.beta(p.k, theta.boot, mu.null, phi.boot, R.boot)
}
\value{
list of diet proportion replicates
}
\description{
Generate bootstrap replicates of diet proportion estimates for
given prey species
}
\keyword{internal}
