% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unbal.diet.data.R
\docType{data}
\name{unbal.diet.data}
\alias{unbal.diet.data}
\title{Sample example of unbalanced repeatability diet estimates data with a max of
two repeated measurements per predator.}
\format{
A data frame with 96 predator diets (50 unique predators) and
13 variables:
\describe{
\item{Seal.ID}{Predator (1 to 50)}
\item{Year}{Either 1 or 2}
\item{capelin}{estimated diet proportion}
\item{coho}{estimated diet proportion}
\item{eulachon}{estimated diet proportion}
\item{herring}{estimated diet proportion}
\item{mackerel}{estimated diet proportion}
\item{pilchard}{estimated diet proportion}
\item{pollock}{estimated diet proportion}
\item{sandlance}{estimated diet proportion}
\item{squid}{estimated diet proportion}
\item{surfsmelt_s}{estimated diet proportion}
\item{sufsmelt_lg}{estimated diet proportion}
}
}
\usage{
unbal.diet.data
}
\description{
Sample example of unbalanced repeatability diet estimates data with a max of
two repeated measurements per predator.
}
\keyword{datasets}
