% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/All_Rep_Code.R
\name{unbal.rho.boot.fun}
\alias{unbal.rho.boot.fun}
\title{RETURNS RHO AND THE BOOTSTRAP SAMPLE DETERMINED BY ind.}
\usage{
unbal.rho.boot.fun(ind, data, gamma.rho)
}
\description{
RETURNS RHO AND THE BOOTSTRAP SAMPLE DETERMINED BY ind.
}
\keyword{internal}
