% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/All_Rep_Code.R
\name{unbal.rho.jack.fun}
\alias{unbal.rho.jack.fun}
\title{FUNCTION TO BE PASSED TO bootstrap::jackknife.}
\usage{
unbal.rho.jack.fun(ind, data, gamma.rho)
}
\description{
FUNCTION TO BE PASSED TO bootstrap::jackknife.
}
\keyword{internal}
