% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Confidence_Interval_Beta_Distribution.R
\name{uniroot.beta}
\alias{uniroot.beta}
\title{Use uniroot() to find roots and compare with bisection outcome}
\usage{
uniroot.beta(x1, x2, alpha, par.list, R, p.mat, k)
}
\description{
Use uniroot() to find roots and compare with bisection outcome
}
\keyword{internal}
