% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/All_Selection_Algorithm_Code.R
\name{zeroEstimates}
\alias{zeroEstimates}
\title{Reintroduce excluded species to diet estimates as forced zeroes}
\usage{
zeroEstimates(estimates, prey.mat)
}
\description{
Reintroduce excluded species to diet estimates as forced zeroes
}
\keyword{internal}
