% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qfasa.R
\name{AIT.obj}
\alias{AIT.obj}
\title{Used in \code{solnp()} as the objective function to be minimized when
Aitchison distance measure is chosen.}
\usage{
AIT.obj(alpha, seal, prey.quantiles)
}
\arguments{
\item{alpha}{vector over which minimization takes place.}

\item{seal}{fatty acid signature of predator.}

\item{prey.quantiles}{matrix of fatty acid signatures of
prey. Each row contains an individual prey signature from a different
species.}
}
\description{
Used in \code{solnp()} as the objective function to be minimized when
Aitchison distance measure is chosen.
}
