% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Confidence_Interval_Beta_Distribution.R
\name{beta.meths.CI}
\alias{beta.meths.CI}
\title{Returns indvidual confidence intervals and simultaneous confidence intervals
based (not bias corrected - see note below) on the zero-inflated beta distribution.}
\usage{
beta.meths.CI(seal.mat, prey.mat, cal.mat, dist.meas, noise, nprey, R.p,
  R.ps, R, p.mat, alpha, FC, ext.fa)
}
\arguments{
\item{seal.mat}{matrix containing the FA signatures of the predators.}

\item{prey.mat}{prey database. A dataframe with first column a
Species label and other columns fatty acid proportions. Fatty
acid proportions are compositional.}

\item{cal.mat}{matrix of calibration coefficients of
predators. Each column corresponds to a different predator. At
least one calibration coefficient vector must be supplied.}

\item{dist.meas}{distance measure to use for estimation: 1=KL,
2=AIT or 3=CS}

\item{noise}{proportion of noise to include in the simulation.}

\item{nprey}{number of prey to sample from the the prey database
when generating pseudo-predators for the nuisance parameter
estimation.}

\item{R.p}{number of beta diet distributions to generate for the
nuisance parameters.}

\item{R.ps}{number of pseudo predators to generate when estimating
nuisance parameters.}

\item{R}{number of bootstrap replicates to use when generating
p-values for confidence interval estimation.}

\item{p.mat}{matrix of predator diet estimates for which we are
trying to find confidence interavls.}

\item{alpha}{confidence interval confidence level.}

\item{FC}{vector of prey fat content. Note that this vector is
passed to the \code{\link{gen.pseudo.seals}} which expects fat
content values for individual prey samples while
\code{\link{pseaudo.seal}} and  \code{\link{p.QFASA}}
expect a species average.}

\item{ext.fa}{subset of FA's to be used to obtain QFASA diet
estimates.}
}
\value{
indvidual confidence intervals and simultaneous confidence
    intervals based on the zero-inflated beta distribution. These
    intervals are biased and should be corrected using the output
    from \code{\link{bias.all}}. \code{ci.l.1} and \code{ci.u.1}
    contain the simultaneous confidence intervals.
}
\description{
For details see:  
    Stewart, C. (2013)  Zero-Inflated Beta Distribution for Modeling the Proportions in 
    Quantitative Fatty Acid Signature Analysis.  
    Journal of Applied Statistics, 40(5), 985-992.
}
\details{
Note:
\itemize{
    \item These intervals are biased and should be corrected using the
          output from \code{\link{bias.all}}.
    \item \code{CI.L.1} and \code{CI.U.1} contain the simultaneous
          confidence intervals.
    \item Slow because of bisection and lots of repetition.
    \item Need to replace \code{p.prey} with
          \code{\link{p.QFASA}} eventually but just use
          p.prey for now. Use example where we estimate a single
          seal diet to compare the estimates from each method.
}
}
\examples{
# Prey
prey.db <- preyFAs \%>\%
    select_(.dots = c('Species', FAset$FA)) \%>\%
    dplyr::arrange(Species)

# Predators
predator.matrix = predatorFAs \%>\% select_(.dots = FAset$FA)

# Diet
diet.est = data.frame(capelin=0.3,
                      coho=0.3,
                      euchalon = 0.01,
                      herring=0.025,
                      mackrel=0.15,
                      pilchard=0.025,
                      pollock=0.025,
                      sandlance=0.15,
                      squid=0.025,
                      surfsmelt_lg = 0.01,
                      surfsmelt_s = 0.01)[rep(1, nrow(predator.matrix)), ]

ci = beta.meths.CI(seal.mat = predator.matrix,
                   prey.mat = prey.db,
                   cal.mat = rep(1, nrow(FAset)),
                   dist.meas = 2,
                   noise = 0,
                   nprey = 50,
                   R.p = 1,
                   R.ps = 10,
                   R = 10, 
                   p.mat = diet.est,
                   alpha = 0.05,
                   FC = rep(1, nrow(prey.db)),
                   ext.fa = FAset$FA)

bias <- bias.all(p.mat = diet.est,
                 prey.mat = prey.db,
                 cal.mat = as.matrix(rep(1, nrow(FAset))),
                 fat.cont = rep(1, nrow(prey.db)),
                 R.bias = 10,
                 noise = 0,
                 nprey = 50,
                 specify.noise = rep(FALSE, nspecies),
                 dist.meas = 2,
                 ext.fa = FAset$FA)

# SIMULTANEOUS CONFIDENCE INTERVALS:
# LOWER LIMIT
ci[[1]] - bias[3,]

# UPPER LIMIT 
ci[[2]] - bias[3,]

}
\references{
Stewart, C. (2013) Zero-inflated beta distribution for
    modeling the proportions in quantitative fatty acid signature
    analysis. Journal of Applied Statistics, 40(5), 985-992.
}
