% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Confidence_Interval_Beta_Distribution.R
\name{bias.all}
\alias{bias.all}
\title{Calculate bias correction for confidence intervals from \code{\link{beta.meths.CI()}}.}
\usage{
bias.all(p.mat, prey.mat, cal.mat, fat.cont, R.bias, noise, nprey,
  specify.noise, dist.meas, ext.fa = ext.common.fa.list)
}
\arguments{
\item{p.mat}{matrix containing the fatty acid signatures of the predators.}

\item{prey.mat}{matrix containing a representative fatty acid signature}

\item{cal.mat}{matrix of calibration factors where the \emph{i} th
column is to be used with the \emph{i} th predator. If modelling is
to be done without calibration coefficients, simply pass a
vector or matrix of ones.}

\item{fat.cont}{prey fat content}

\item{R.bias}{botstrap replicates}

\item{noise}{noise}

\item{nprey}{number of prey}

\item{specify.noise}{noise}

\item{dist.meas}{distance measure}

\item{ext.fa}{subset of FA's to use.}
}
\value{
Row 1 is Lambda CI, row 2 is Lambda skew, and row 3 is Beta CI
}
\description{
Calculate bias correction for confidence intervals from \code{\link{beta.meths.CI()}}.
}
