% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qfasa.R
\name{chisq.CA}
\alias{chisq.CA}
\title{chisq.CA
CALLED BY create.d.mat, computes the chisquare distance similar to the one discussed
in (PAWLOWSKY-GLAHN AND BUCCIANTI, 2011)}
\usage{
chisq.CA(x1, x2)
}
\arguments{
\item{x1}{TODO}

\item{x2}{TODO}
}
\description{
chisq.CA
CALLED BY create.d.mat, computes the chisquare distance similar to the one discussed
in (PAWLOWSKY-GLAHN AND BUCCIANTI, 2011)
}
