% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qfasa.R
\name{chisq.CA}
\alias{chisq.CA}
\title{Called by \code{create.d.mat()} to compute the chi-square distance.}
\usage{
chisq.CA(x1, x2)
}
\arguments{
\item{x1}{vector}

\item{x2}{vector}
}
\description{
Called by \code{create.d.mat()} to compute the chi-square distance.
}
