% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qfasa.R
\name{chisq.dist}
\alias{chisq.dist}
\title{Returns the distance between two compositional vectors using the chi-square distance.}
\usage{
chisq.dist(x.1, x.2, gamma)
}
\arguments{
\item{x.1}{compositional vector}

\item{x.2}{compositional vector}

\item{gamma}{power transform taken to be 1.}
}
\description{
Returns the distance between two compositional vectors using the chi-square distance.
}
\references{
Stewart, C., Iverson, S. and Field, C. (2014) Testing for a change in
diet using fatty acid signatures.  Environmental and Ecological Statistics 21, pp. 775-792.

Connie Stewart (2017) An approach to measure distance between compositional diet estimates containing essential zeros, Journal of Applied Statistics, 44:7, 1137-1152, DOI: 10.1080/02664763.2016.1193846
}
