% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qfasa.R
\name{create.d.mat}
\alias{create.d.mat}
\title{Called by testfordiff.ind.boot.fun to create a matrix of distances.}
\usage{
create.d.mat(Y.1, Y.2)
}
\arguments{
\item{Y.1}{TODO}

\item{Y.2}{TODO}
}
\description{
Called by testfordiff.ind.boot.fun to create a matrix of distances.
}
