% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qfasa.R
\name{create.d.mat}
\alias{create.d.mat}
\title{create.d.mat
CALLED BY testfordiff.ind.boot.fun, want to create a matrix of distances 
d11!=0 and d12!=d21 since not the same #1 and 1 are not the same seals.}
\usage{
create.d.mat(Y.1, Y.2)
}
\arguments{
\item{Y.1}{TODO}

\item{Y.2}{TODO}
}
\description{
create.d.mat
CALLED BY testfordiff.ind.boot.fun, want to create a matrix of distances 
d11!=0 and d12!=d21 since not the same #1 and 1 are not the same seals.
}
