% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qfasa.R
\name{KL.dist}
\alias{KL.dist}
\title{Returns the distance between two compositional vectors using Kullback-Leibler
distance measure.}
\usage{
KL.dist(x.1, x.2)
}
\arguments{
\item{x.1}{compositional vector}

\item{x.2}{compositional vector}
}
\description{
Returns the distance between two compositional vectors using Kullback-Leibler
distance measure.
}
\references{
S.J. Iverson, C. Field, W.D. Bowen, and W. Blanchard (2004)  Quantitative
fatty acid signature analysis: A new method of estimating predator diets, Ecological
Monographs 72, pp. 211-235.
}

