% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qfasa.R
\name{KL.obj}
\alias{KL.obj}
\title{Used in \code{solnp()} as the objective function to be minimized when
Kullback-Leibler distance measure is chosen.}
\usage{
KL.obj(alpha, predator, prey.quantiles)
}
\arguments{
\item{alpha}{vector over which minimization takes place.}

\item{predator}{fatty acid signature of predator.}

\item{prey.quantiles}{matrix of fatty acid signatures of
prey. Each row contains an individual prey signature from a different
species.}
}
\description{
Used in \code{solnp()} as the objective function to be minimized when
Kullback-Leibler distance measure is chosen.
}
