% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/All_Simplified_MUFASA_Code.R
\name{multiplicativeReplacement}
\alias{multiplicativeReplacement}
\title{Multiplicative replacement of zeroes}
\usage{
multiplicativeReplacement(compositional.mat, delta = 1e-05)
}
\description{
Multiplicative replacement of zeroes
}
\keyword{internal}
