% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qfasa.R
\name{p.QFASA}
\alias{p.QFASA}
\title{Computes the diet estimate for each predator in seal.mat using
either the Kullback-Leibler Distance (KL), the Aitchison Distance
(AIT) or the Chi-Square Distance (CS).}
\usage{
p.QFASA(seal.mat, prey.mat, cal.mat, dist.meas, gamma, FC = rep(1,
  nrow(prey.mat)), start.val = rep(0.99999, nrow(prey.mat)), ext.fa)
}
\arguments{
\item{seal.mat}{matrix containing the FA signatures of the predators.}

\item{prey.mat}{matrix containing a representative FA signature
from each prey group (usually the mean). The first column must
index the prey group.}

\item{cal.mat}{matrix of calibration factors where the \emph{i} th
column is to be used with the \emph{i} th seal}

\item{dist.meas}{distance measure to use for estimation: 1=KL,
2=AIT or 3=CS}

\item{gamma}{parameter required for calculations using CS distance
(passed to CS.obj). Currently being set to 1.}

\item{FC}{vector of fat content}

\item{start.val}{initial vector of parameters to be optimized}

\item{ext.fa}{subset of FA's to be used to obtain QFASA diet estimates.}
}
\value{
a matrix of the diet estimate for each predator by the modelling groups and a list of 
lists where each list (one per predator) is itself a list of four outputs: 
\item{ModFAS}{the value of the modelled FA (i.e. after CCs have been applied and the FA subsetted and renormalised over the designated FA set).
 These are expressed as proportions (they will sum to 1).}
\item{DistCont}{the contribution of each FA to the final minimized distance.} 
\item{PropDistCont}{the contribution of each FA to the final minimized distance as a proportion of the total.}
\item{MinDist}{ the final minimized distance in the code below the 'ldply' function from the plyr package is used 
to compile the lists within 'Additional Measures' into a data frame with one row per predator (in the same order 
as the input predator matrix) and the values for each of the 4 lists arranged into columns. The 'ldply' function 
automatically names the columns of the data frame with a concatenation of the originating list name and the FA name
 so that the 4 sets of outputs can be easily identified within the data frame.}
}
\description{
Computes the diet estimate for each predator in seal.mat using
either the Kullback-Leibler Distance (KL), the Aitchison Distance
(AIT) or the Chi-Square Distance (CS).
}
\examples{
 ## Fatty Acids
 data(FAset)
 fa.set = as.vector(unlist(FAset))
 
 ## Predators
 data(predatorFAs)
 tombstone.info = predatorFAs[,1:4]
 predator.matrix = predatorFAs[,5:(ncol(predatorFAs))]
 npredators = nrow(predator.matrix)

 ## Prey
 data(preyFAs)
 prey.sub=(preyFAs[,4:(ncol(preyFAs))])[fa.set]
 prey.sub=prey.sub/apply(prey.sub,1,sum) 
 group=as.vector(preyFAs$Species)
 prey.matrix=cbind(group,prey.sub)
 prey.matrix=MEANmeth(prey.matrix) 

 FC = preyFAs[,c(2,3)] 
 FC = as.vector(tapply(FC$lipid,FC$Species,mean,na.rm=TRUE))

 ## Calibration Coefficients
 data(CC)
 cal.vec = CC[,2]
 cal.mat = replicate(npredators, cal.vec)

 # Run QFASA
 Q = p.QFASA(predator.matrix,
             prey.matrix,
             cal.mat,
             dist.meas=1, 
             gamma=1,
             FC,
             start.val = rep(1,nrow(prey.matrix)),
             fa.set)
 
}
