% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Confidence_Interval_Beta_Distribution.R
\name{pseudo.seal}
\alias{pseudo.seal}
\title{Generates pseudo-seals. Allows:}
\usage{
pseudo.seal(prey.sim, diet, noise, nprey, cal, fat.cont, specify.noise)
}
\arguments{
\item{prey.sim}{OUTPUT OF split.prey}

\item{diet}{DIET COMPOSITION VECTOR  (NOTE: THIS VECTOR SHOULD SUM
TO 1-NOISE.  THE NOISE WILL BE ADDED TO THE diet VECTOR.)}

\item{noise}{AMOUNT OF NOISE}

\item{nprey}{nprey TOTAL NUMBER OF PREY TO BE SAMPLED}

\item{cal}{CALIBRATION FACTORS}

\item{fat.cont}{VECTOR OF FAT CONTENT OF LENGTH=I (# OF SPECIES)}

\item{specify.noise}{A BOOLEAN VECTOR WITH TRUES DENOTING SPECIES
TO USE IN NOISE.}
}
\value{
seal.star SIMULATED SEAL FA SIGNATURE.
}
\description{
1) Fat content to be included in the generated seals and
2) Some species to be truly zero (that is,"zero species" do not
have to be included in the "noise")
}
\details{
Note:  it is assumed that sum(diet) is 1-noise
}
