% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qfasa.R
\name{QFASA.const.eqn}
\alias{QFASA.const.eqn}
\title{Returns \code{sum(alpha)} and used in \code{solnp()}.}
\usage{
QFASA.const.eqn(alpha, predator, prey.quantiles, gamma)
}
\arguments{
\item{alpha}{vector over which minimization takes place.}

\item{predator}{fatty acid signature of predator.}

\item{prey.quantiles}{matrix of fatty acid signatures of
prey. Each row contains an individual prey signature from a different
species.}

\item{gamma}{power transform exponent (see chisq.dist).}
}
\description{
Returns \code{sum(alpha)} and used in \code{solnp()}.
}
