% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/All_CI_Code.R,
%   R/Confidence_Interval_Beta_Distribution.R
\name{split.prey}
\alias{split.prey}
\title{Splits prey database into a simulation set (1/3) and a modelling set (2/3).
Returns a list:}
\usage{
\method{split}{prey}(prey.mat)

\method{split}{prey}(prey.mat)
}
\arguments{
\item{prey.mat}{matrix of individual prey fatty acid signatures
where the first column denotes the prey type}
}
\description{
1. simulation prey database
2. modelling prey database

1. simulation prey database
2. modelling prey database
}
\details{
IF number of samples of a prey type <=5, then prey.mod AND prey.sim
are duplicated instead of split.

IF number of samples of a prey type <=5, then prey.mod AND prey.sim
are duplicated instead of split.
}
