% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Confidence_Interval_Beta_Distribution.R
\name{split.prey}
\alias{split.prey}
\title{Splits prey database into a simulation set (1/3) and a modelling
set (2/3). If number of samples of a prey type is less than or equal to 5,
then \code{prey.mod} and \code{prey.sim} are duplicated instead of split.}
\usage{
\method{split}{prey}(prey.mat)
}
\arguments{
\item{prey.mat}{matrix of individual prey fatty acid signatures
where the first column denotes the prey type}
}
\value{
list of simulation prey database and modelling prey database.
}
\description{
Splits prey database into a simulation set (1/3) and a modelling
set (2/3). If number of samples of a prey type is less than or equal to 5,
then \code{prey.mod} and \code{prey.sim} are duplicated instead of split.
}
