% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qfasa.R
\name{testfordiff.ind.boot.fun}
\alias{testfordiff.ind.boot.fun}
\title{testfordiff.ind.boot.fun
CALLED BY testfordiff.ind.boot}
\usage{
testfordiff.ind.boot.fun(data, i, ns1, change.zero = 1e-05)
}
\arguments{
\item{data}{SAMPLE OF COMPOSITIONAL DATA}

\item{ns1}{SAMPLE SIZE OF seals.1}
}
\description{
testfordiff.ind.boot.fun
CALLED BY testfordiff.ind.boot
}
