% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qfasa.R
\name{testfordiff.ind.boot.fun}
\alias{testfordiff.ind.boot.fun}
\title{Called by testfordiff.ind.boot}
\usage{
testfordiff.ind.boot.fun(data, i, ns1, change.zero = 1e-05)
}
\arguments{
\item{data}{sample of compositional data}

\item{ns1}{sample size of seals.1}
}
\description{
Called by testfordiff.ind.boot
}
