% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qfasa.R
\name{testfordiff.ind.boot}
\alias{testfordiff.ind.boot}
\title{Called by \code{testfordiff.ind.pval()}.}
\usage{
testfordiff.ind.boot(data, ns1, R)
}
\arguments{
\item{data}{sample of compositional data}

\item{ns1}{sample size of compdata.1}

\item{R}{number of bootstrap samples.  default is 500.}
}
\description{
Called by \code{testfordiff.ind.pval()}.
}
