% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qfasa.R
\name{testfordiff.ind.boot}
\alias{testfordiff.ind.boot}
\title{testfordiff.ind.boot
CALLED BY testfordiff.ind.pval, resamples without replacent pooled samples of seals}
\usage{
testfordiff.ind.boot(data, ns1, R)
}
\arguments{
\item{data}{sample of compositional data}

\item{ns1}{sample size of seals.1}

\item{R}{number of bootstrap samples.  default is 500.}
}
\description{
testfordiff.ind.boot
CALLED BY testfordiff.ind.pval, resamples without replacent pooled samples of seals
}
