% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qfasa.R
\name{testfordiff.ind.pval}
\alias{testfordiff.ind.pval}
\title{Test for a difference between two independent samples of compositional data.
Zeros of any type are allowed.}
\usage{
testfordiff.ind.pval(compdata.1, compdata.2, ns1, R = 500)
}
\arguments{
\item{compdata.1}{sample of compositional data.}

\item{compdata.2}{sample of compositional data.}

\item{ns1}{sample size of compdata.1.}

\item{R}{number of bootstrap samples, default is 500.}
}
\value{
p-value obtained through a multivariate permutation test with test statistic based
on chi-square distances.
}
\description{
Test for a difference between two independent samples of compositional data.
Zeros of any type are allowed.
}
\examples{
library(boot)
## Fatty Acids
data(FAset)
fa.set = as.vector(unlist(FAset))

## Prey
data(preyFAs)
prey.sub=(preyFAs[,4:(ncol(preyFAs))])[fa.set]
prey.sub=prey.sub/apply(prey.sub,1,sum)
group=as.vector(preyFAs$Species)
prey.matrix=cbind(group,prey.sub)

## Capelin FA sig
capelin.sig <- prey.matrix[prey.matrix$group=="capelin",]
## Sandlance FA sig
sandlance.si <-prey.matrix[prey.matrix$group=="sandlance",]

## Run testfordiff.ind.pval
testfordiff.ind.pval(as.matrix(capelin.sig[,-1]),as.matrix(sandlance.si[,-1]),nrow(capelin.sig))

}
\references{
Stewart, C., Iverson, S. and Field, C. (2014) Testing for a change in
diet using fatty acid signatures.  Environmental and Ecological Statistics 21, pp. 775-792.
}
