% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qfasa.R
\name{testfordiff.ind.pval}
\alias{testfordiff.ind.pval}
\title{testfordiff.ind.pval}
\usage{
testfordiff.ind.pval(seals.1, seals.2, ns1, R = 500)
}
\arguments{
\item{seals.1}{SAMPLE OF COMPOSITIONAL DATA}

\item{seals.2}{SAMPLE OF COMPOSITIONAL DATA}

\item{ns1}{SAMPLE SIZE OF seals.1}

\item{R}{NUMBER OF BOOTSTRAP SAMPLES.  DEFAULT IS 500.}
}
\description{
testfordiff.ind.pval
}
\references{
Stewart, C., Iverson, S. and Field, C. (2014) Testing for a change in
diet using fatty acid signatures.  Environmental and Ecological Statistics 21, pp. 775-792.
}
